pyproject.toml
[project]

# General Project Information
# -------------------------------------------------------------------------------------------------------------------
name = "arxiv_paper_hound"
version = "0.1.0"
requires-python = ">= 3.10"
description = "A semantic search engine for finding ArXiv academic papers based on query meaning, not just keywords."
readme = "README.md"
authors = [{name = "Gordeev A.", email = "asegordeev@gmail.com"}]
maintainers = [{name = "Gordeev A.", email = "asegordeev@gmail.com"}]
classifiers=[
    # --- Project Status ---
    'Development Status :: 3 - Alpha',
    # 'Development Status :: 4 - Beta',
    # 'Development Status :: 5 - Production/Stable',

    # --- Intended Audience ---
    'Intended Audience :: Science/Research',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',

    # --- License ---
    'License :: OSI Approved :: MIT License',

    # --- Topic / Subject ---
    'Topic :: Scientific/Engineering',
    'Topic :: Scientific/Engineering :: Information Analysis',
    'Topic :: Text Processing',
    'Topic :: Text Processing :: Indexing',
    'Topic :: Internet :: WWW/HTTP :: Indexing/Search',

    # --- Supported Python Versions ---
    'Programming Language :: Python :: 3',

    # --- Environment & OS ---
    'Operating System :: OS Independent',
    'Natural Language :: English',
    'Typing :: Typed',
]
license = {text = "MIT"}
# -------------------------------------------------------------------------------------------------------------------

# Core Dependencies
dependencies = [
    "arxiv==2.2.0",
    "dependency-injector>=4.48.2",
    "fastapi==0.116.2",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "qdrant-client==1.15.1",
    "sentence-transformers==5.0.0",
    "starlette-context>=0.4.0",
    "tqdm>=4.67.1",
    "uvicorn==0.35.0",
]

[project.optional-dependencies]
# Optional dependencies for development
dev = [
    "coverage==7.10.6",  # testing
    "pytest==8.4.2",  # testing
    "ruff==0.13.0",  # linting
    "ty==0.0.1a20", # checking types
]

# Linters and Formatters
# -------------------------------------------------------------------------------------------------------------------
[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.
# -------------------------------------------------------------------------------------------------------------------

# -------------------------------------------------------------------------------------------------------------------
[tool.ruff]
indent-width = 4
line-length = 120
output-format = "grouped"
respect-gitignore = true
extend-exclude = [
    "docs/",
]
show-fixes = true

[tool.ruff.format]
line-ending = "lf"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "YTT", # flake8-2020
    "CPY", # flake8-copyright
    "FA", # flake8-future-annotations
    "TD", # flake8-todos
    "C90", # mccabe
    "PGH", # pygrep-hooks

    # disable these rules to use the ruff formatter.
    "COM819", # prohibited-trailing-comma
    "D206",   # docstring-tab-indentation
    "D300",   # triple-single-quotes
    "E111",   # indentation-with-invalid-multiple
    "E114",   # indentation-with-invalid-multiple-comment
    "E117",   # over-indented
    "Q000",   # bad-quotes-inline-string
    "Q001",   # bad-quotes-multiline-string
    "Q002",   # bad-quotes-docstring
    "Q003",   # avoidable-escaped-quote
    "W191",   # tab-indentation
]
task-tags = ["TODO", "FIXME", "XXX", "HACK"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]   # ignore unused imports in __init__.py files
"test_*.py" = ["ANN"]      # Ignore missing type annotations in tests

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true  # suppress ANN401 for *args and **kwargs
# -------------------------------------------------------------------------------------------------------------------
